Use timeslots from Redux store in Answer page
